+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/netty/MockServer.java
@@ -20,14 +20,20 @@ import de.gematik.test.tiger.mockserver.mock.action.http.HttpActionHandler;
 import de.gematik.test.tiger.mockserver.model.HttpRequest;
 import de.gematik.test.tiger.mockserver.proxyconfiguration.ProxyConfiguration;
 import de.gematik.test.tiger.mockserver.socket.tls.NettySslContextFactory;
+import de.gematik.test.tiger.proxy.data.TigerConnectionStatus;
 import io.netty.bootstrap.ServerBootstrap;
 import io.netty.buffer.PooledByteBufAllocator;
 import io.netty.channel.ChannelOption;
 import io.netty.channel.WriteBufferWaterMark;
 import io.netty.channel.socket.nio.NioServerSocketChannel;
 import java.net.InetSocketAddress;
+import java.net.SocketAddress;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.stream.Collectors;
 import javax.annotation.Nullable;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
@@ -41,6 +47,7 @@ public class MockServer extends LifeCycle {
 
   private InetSocketAddress remoteSocket;
   private HttpActionHandler actionHandler;
+  private Map<SocketAddress, TigerConnectionStatus> connectionStatusMap = new ConcurrentHashMap<>();
 
   /**
    * Start the instance using the ports provided
@@ -204,6 +211,7 @@ public class MockServer extends LifeCycle {
             .option(
                 ChannelOption.WRITE_BUFFER_WATER_MARK,
                 new WriteBufferWaterMark(8 * 1024, 32 * 1024))
+            .childHandler(new ConnectionCounterHandler(this))
             .childHandler(
                 new MockServerUnificationInitializer(
                     configuration,
@@ -228,15 +236,12 @@ public class MockServer extends LifeCycle {
     return remoteSocket;
   }
 
-  public ExpectationBuilder when(
-      HttpRequest httpRequest, Integer priority) {
-    return new ExpectationBuilder(
-        new Expectation(httpRequest, priority), this);
+  public ExpectationBuilder when(HttpRequest httpRequest, Integer priority) {
+    return new ExpectationBuilder(new Expectation(httpRequest, priority), this);
   }
 
   public ExpectationBuilder when(HttpRequest requestDefinition) {
-    return new ExpectationBuilder(
-        new Expectation(requestDefinition, 0), this);
+    return new ExpectationBuilder(new Expectation(requestDefinition, 0), this);
   }
 
   public void removeExpectation(String expectationId) {
@@ -245,5 +250,18 @@ public class MockServer extends LifeCycle {
 
   public List<Expectation> retrieveActiveExpectations() {
     return httpState.retrieveActiveExpectations();
+  }
+
+  public synchronized void addConnectionWithStatus(SocketAddress socketAddress, TigerConnectionStatus status) {
+    connectionStatusMap.put(socketAddress, status);
+  }
+
+  public synchronized Map<SocketAddress, TigerConnectionStatus> getOpenConnections() {
+    return connectionStatusMap.entrySet().stream()
+      .collect(Collectors.toMap(Entry::getKey, Entry::getValue));
+  }
+
+  public synchronized void removeRemoteAddress(SocketAddress socketAddress) {
+    connectionStatusMap.remove(socketAddress);
   }
 }