+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/netty/MockServerUnificationInitializer.java
@@ -5,7 +5,6 @@
 package de.gematik.test.tiger.mockserver.netty;
 
 import de.gematik.test.tiger.mockserver.configuration.Configuration;
-import de.gematik.test.tiger.mockserver.lifecycle.LifeCycle;
 import de.gematik.test.tiger.mockserver.mock.HttpState;
 import de.gematik.test.tiger.mockserver.mock.action.http.HttpActionHandler;
 import de.gematik.test.tiger.mockserver.netty.unification.PortUnificationHandler;
@@ -13,34 +12,26 @@ import de.gematik.test.tiger.mockserver.socket.tls.NettySslContextFactory;
 import io.netty.channel.ChannelHandler;
 import io.netty.channel.ChannelHandlerAdapter;
 import io.netty.channel.ChannelHandlerContext;
+import lombok.Data;
+import lombok.extern.slf4j.Slf4j;
 
 /*
  * @author jamesdbloom
  */
 @ChannelHandler.Sharable
+@Data
+@Slf4j
 public class MockServerUnificationInitializer extends ChannelHandlerAdapter {
   private final Configuration configuration;
-  private final LifeCycle server;
+  private final MockServer server;
   private final HttpState httpState;
   private final HttpActionHandler actionHandler;
   private final NettySslContextFactory nettySslContextFactory;
 
-  public MockServerUnificationInitializer(
-      Configuration configuration,
-      LifeCycle server,
-      HttpState httpState,
-      HttpActionHandler actionHandler,
-      NettySslContextFactory nettySslContextFactory) {
-    this.configuration = configuration;
-    this.server = server;
-    this.httpState = httpState;
-    this.actionHandler = actionHandler;
-    this.nettySslContextFactory = nettySslContextFactory;
-  }
-
   @Override
   public void handlerAdded(ChannelHandlerContext ctx) {
     ctx.pipeline()
+        .addFirst(new ConnectionCounterHandler(server))
         .replace(
             this,
             null,