+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/netty/unification/PortUnificationHandler.java
@@ -15,15 +15,16 @@ import static java.util.Collections.unmodifiableSet;
 
 import de.gematik.test.tiger.mockserver.codec.MockServerHttpServerCodec;
 import de.gematik.test.tiger.mockserver.configuration.Configuration;
-import de.gematik.test.tiger.mockserver.lifecycle.LifeCycle;
 import de.gematik.test.tiger.mockserver.mappers.MockServerHttpResponseToFullHttpResponse;
 import de.gematik.test.tiger.mockserver.mock.HttpState;
 import de.gematik.test.tiger.mockserver.mock.action.http.HttpActionHandler;
 import de.gematik.test.tiger.mockserver.model.HttpResponse;
 import de.gematik.test.tiger.mockserver.netty.HttpRequestHandler;
+import de.gematik.test.tiger.mockserver.netty.MockServer;
 import de.gematik.test.tiger.mockserver.netty.proxy.BinaryHandler;
 import de.gematik.test.tiger.mockserver.socket.tls.NettySslContextFactory;
 import de.gematik.test.tiger.mockserver.socket.tls.SniHandler;
+import de.gematik.test.tiger.proxy.data.TigerConnectionStatus;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
 import io.netty.channel.*;
@@ -58,7 +59,7 @@ public class PortUnificationHandler extends ReplayingDecoder<Void> {
       new ConcurrentHashMap<>();
   private final HttpContentLengthRemover httpContentLengthRemover = new HttpContentLengthRemover();
   private final Configuration configuration;
-  private final LifeCycle server;
+  private final MockServer server;
   private final HttpState httpState;
   private final HttpActionHandler actionHandler;
   private final NettySslContextFactory nettySslContextFactory;
@@ -66,7 +67,7 @@ public class PortUnificationHandler extends ReplayingDecoder<Void> {
 
   public PortUnificationHandler(
       Configuration configuration,
-      LifeCycle server,
+      MockServer server,
       HttpState httpState,
       HttpActionHandler actionHandler,
       NettySslContextFactory nettySslContextFactory) {
@@ -98,14 +99,6 @@ public class PortUnificationHandler extends ReplayingDecoder<Void> {
     } else {
       return false;
     }
-  }
-
-  public static void enableSslDownstream(Channel channel) {
-    channel.attr(TLS_ENABLED_DOWNSTREAM).set(Boolean.TRUE);
-  }
-
-  public static void disableSslDownstream(Channel channel) {
-    channel.attr(TLS_ENABLED_DOWNSTREAM).set(Boolean.FALSE);
   }
 
   public static boolean isSslEnabledDownstream(Channel channel) {
@@ -147,6 +140,7 @@ public class PortUnificationHandler extends ReplayingDecoder<Void> {
 
   private void enableTls(ChannelHandlerContext ctx, ByteBuf msg) {
     ChannelPipeline pipeline = ctx.pipeline();
+    server.addConnectionWithStatus(ctx.channel().remoteAddress(), TigerConnectionStatus.OPEN_TLS);
     pipeline.addFirst(new SniHandler(configuration, nettySslContextFactory));
     enableSslUpstreamAndDownstream(ctx.channel());